Add 'never.ts' to demonstrate use of 'never' type
Added a new file, 'never.ts', that showcases the use of the TypeScript 'never' type. The file defines a function that throws an error, thus not returning a value and therefore having the return type of 'never'. It also demonstrates how to assign a value to a variable of 'never' type.